# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.CreateScheduleRequest:
    properties:
      arrival_station:
        type: string
      arrival_time:
        type: string
      available_seats:
        minimum: 0
        type: integer
      departure_station:
        type: string
      departure_time:
        type: string
      price:
        minimum: 0
        type: number
      train_id:
        type: integer
    required:
    - arrival_station
    - arrival_time
    - available_seats
    - departure_station
    - departure_time
    - price
    - train_id
    type: object
...
  dto.CreateUserRequest:
    properties:
      email:
        type: string
      full_name:
        type: string
      password:
        minLength: 6
        type: string
      phone:
        type: string
      role:
        enum:
        - admin
        - user
        type: string
    required:
    - email
    - full_name
    - password
    - role
    type: object
  dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.LoginResponse:
    properties:
      token:
        type: string
      user: {}
    type: object
  dto.RegisterRequest:
    properties:
      email:
        type: string
      full_name:
        type: string
      password:
        minLength: 6
        type: string
      phone:
        type: string
    required:
    - email
    - full_name
    - password
    - phone
    type: object
...
  models.Schedule:
    properties:
      arrival_station:
        type: string
      arrival_time:
        type: string
      available_seats:
        type: integer
      created_at:
        type: string
      departure_station:
        type: string
      departure_time:
        type: string
      id:
        type: integer
      modified_at:
        type: string
      price:
        type: number
      train_id:
        type: integer
      trainCode:
        type: string
      trainName:
        type: string
      trainType:
        type: string
    type: object
...
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: integer
      modified_at:
        type: string
      phone:
        type: string
      role:
        type: string
    type: object
  utils.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API Server untuk aplikasi pemesanan tiket kereta api TiketSepur
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: TiketSepur API
  version: "1.0"
paths:
  /admin/schedules:
    get:
      consumes:
      - application/json
      description: Get all train schedules (admin only)
      produces:
      - application/json
      responses:
        "200":
          description: List of schedules
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Schedule'
                  type: array
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Get all schedules
      tags:
      - schedules
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT token
      parameters:
      - description: Login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.LoginResponse'
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Authentication failed
          schema:
            $ref: '#/definitions/utils.Response'
      summary: User login
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Logout user and invalidate token
      produces:
      - application/json
      responses:
        "200":
          description: Logout successful
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Logout failed
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: User logout
      tags:
      - auth
  /auth/me:
    get:
      consumes:
      - application/json
      description: Get information about the currently logged-in user
      produces:
      - application/json
      responses:
        "200":
          description: User info retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Get current user info
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user account
      parameters:
      - description: User registration details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User registered successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.User'
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Register new user
      tags:
      - auth
  /auth/register-admin:
    post:
      consumes:
      - application/json
      description: Register a new admin user account
      parameters:
      - description: Admin registration details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Admin registered successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.User'
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Register admin user
      tags:
      - auth
  /payments/confirm/{paymentCode}:
    post:
      consumes:
      - application/json
      description: Confirm a payment for a ticket booking
      parameters:
      - description: Payment Code
        in: path
        name: paymentCode
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Payment confirmed
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: Invalid payment code
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Confirm payment
      tags:
      - payments
  /payments/status/{paymentCode}:
    get:
      consumes:
      - application/json
      description: Get the current status of a payment
      parameters:
      - description: Payment Code
        in: path
        name: paymentCode
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Payment status
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: Invalid payment code
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Get payment status
      tags:
      - payments
  /schedules:
    get:
      consumes:
      - application/json
      description: Get all available train schedules for public
      produces:
      - application/json
      responses:
        "200":
          description: List of schedules
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Schedule'
                  type: array
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get all public schedules
      tags:
      - schedules
    post:
      consumes:
      - application/json
      description: Create a new train schedule
      parameters:
      - description: Schedule details
        in: body
        name: schedule
        required: true
        schema:
          $ref: '#/definitions/dto.CreateScheduleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Schedule created successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Schedule'
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Create a new schedule
      tags:
      - schedules
  /schedules/{id}:
    get:
      consumes:
      - application/json
      description: Get train schedule by its ID
      parameters:
      - description: Schedule ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Schedule details
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Schedule'
              type: object
        "404":
          description: Schedule not found
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get schedule by ID
      tags:
      - schedules
  /schedules/search:
    get:
      consumes:
      - application/json
      description: Search train schedules by criteria
      parameters:
      - description: Origin station
        in: query
        name: origin
        type: string
      - description: Destination station
        in: query
        name: destination
        type: string
      - description: Travel date (YYYY-MM-DD)
        in: query
        name: date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of matching schedules
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Schedule'
                  type: array
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Search schedules
      tags:
      - schedules
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"